interface fixes
