+++ src/components/layout/CippPage.js
@@ -16,45 +16,13 @@ export function CippPage({
   titleButton = null,
   wizard = false,
 }) {
-  const { data: tenants = [], isSuccess } = useListTenantsQuery({ showAllTenantSelector })
-  const tenant = useSelector((state) => state.app.currentTenant)
-  const [searchParams, setSearchParams] = useSearchParams()
-  const navigate = useNavigate()
-  const dispatch = useDispatch()
-
-  const updateSearchParams = useCallback(
-    (params) => {
-      navigate(`${queryString(params)}`, { replace: true })
-    },
-    [navigate],
-  )
-
-  useEffect(() => {
-    if (tenantSelector) {
-      const customerId = searchParams.get('customerId')
-      if (customerId && isSuccess) {
-        const currentTenant = tenants.filter((tenant) => tenant.customerId === customerId)
-        if (currentTenant.length > 0) {
-          dispatch(setCurrentTenant({ tenant: currentTenant[0] }))
-        }
-      }
-      if (!customerId && Object.keys(tenant).length > 0) {
-        updateSearchParams({ customerId: tenant?.customerId })
-      }
-    }
-  }, [dispatch, isSuccess, searchParams, tenant, tenantSelector, tenants, updateSearchParams])
-
-  const handleTenantSelect = (tenant) => {
-    setSearchParams({ customerId: tenant.customerId })
-  }
-
   return (
     <div>
       <>
         {wizard && (
           <CCard className="content-card">
             <CCardBody>
-              {tenantSelector && Object.keys(tenant).length === 0 ? (
+              {tenantSelector ? (
                 <CCallout className="mb-0" color="warning">
                   Select a tenant to get started.
                 </CCallout>
@@ -66,7 +34,7 @@ export function CippPage({
         )}
         {!wizard && (
           <CCardBody>
-            {tenantSelector && Object.keys(tenant).length === 0 ? (
+            {tenantSelector ? (
               <CCallout className="mb-0" color="warning">
                 Select a tenant to get started.
               </CCallout>
