+++ src/scss/_custom.scss
@@ -553,3 +553,6 @@
   margin-top: 69px;
   height: calc(100vh - 69px);
 }
+.wrapper.d-flex.flex-column.min-vh-100 .card {
+  margin-bottom: 0px;
+}
