+++ src/views/email-exchange/transport/ListTransportTemplates.js
@@ -2,7 +2,15 @@ import React, { useState } from 'react'
 import { useSelector } from 'react-redux'
 import { CippCodeBlock, CippOffcanvas } from 'src/components/utilities'
 import { CippDatatable } from 'src/components/tables'
-import { CCardBody, CButton, CCallout, CSpinner } from '@coreui/react'
+import {
+  CCardBody,
+  CButton,
+  CCallout,
+  CSpinner,
+  CCard,
+  CCardHeader,
+  CCardTitle,
+} from '@coreui/react'
 import { faEye, faTrash } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useLazyGenericGetRequestQuery } from 'src/store/api/app'
@@ -83,24 +91,29 @@ const TransportListTemplates = () => {
 
   return (
     <CippPage title="Available Transport Rule Templates" tenantSelector={false}>
-      <CCardBody>
-        {getResults.isFetching && (
-          <CCallout color="info">
-            <CSpinner>Loading</CSpinner>
-          </CCallout>
-        )}
-        {getResults.isSuccess && <CCallout color="info">{getResults.data?.Results}</CCallout>}
-        {getResults.isError && (
-          <CCallout color="danger">Could not connect to API: {getResults.error.message}</CCallout>
-        )}
-        <CippDatatable
-          keyField="id"
-          reportName={`${tenant?.defaultDomainName}-ListTransportRulesTemplates-List`}
-          path="/api/ListTransportRulesTemplates"
-          columns={columns}
-          params={{ TenantFilter: tenant?.defaultDomainName }}
-        />
-      </CCardBody>
+      <CCard className="content-card">
+        <CCardHeader className="d-flex justify-content-between align-items-center">
+          <CCardTitle>Results</CCardTitle>
+        </CCardHeader>
+        <CCardBody>
+          {getResults.isFetching && (
+            <CCallout color="info">
+              <CSpinner>Loading</CSpinner>
+            </CCallout>
+          )}
+          {getResults.isSuccess && <CCallout color="info">{getResults.data?.Results}</CCallout>}
+          {getResults.isError && (
+            <CCallout color="danger">Could not connect to API: {getResults.error.message}</CCallout>
+          )}
+          <CippDatatable
+            keyField="id"
+            reportName={`${tenant?.defaultDomainName}-ListTransportRulesTemplates-List`}
+            path="/api/ListTransportRulesTemplates"
+            columns={columns}
+            params={{ TenantFilter: tenant?.defaultDomainName }}
+          />
+        </CCardBody>
+      </CCard>
     </CippPage>
   )
 }
