+++ src/views/endpoint/MEM/MEMListPolicyTemplates.js
@@ -2,7 +2,15 @@ import React, { useState } from 'react'
 import { useSelector } from 'react-redux'
 import { CippCodeBlock, CippOffcanvas } from 'src/components/utilities'
 import { CippDatatable } from 'src/components/tables'
-import { CCardBody, CButton, CCallout, CSpinner } from '@coreui/react'
+import {
+  CCardBody,
+  CButton,
+  CCallout,
+  CSpinner,
+  CCardHeader,
+  CCardTitle,
+  CCard,
+} from '@coreui/react'
 import { faEye, faTrash } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useLazyGenericGetRequestQuery } from 'src/store/api/app'
@@ -91,24 +99,30 @@ const AutopilotListTemplates = () => {
 
   return (
     <CippPage title="Available Endpoint Manager Templates" tenantSelector={false}>
-      <CCardBody>
-        {getResults.isFetching && (
-          <CCallout color="info">
-            <CSpinner>Loading</CSpinner>
-          </CCallout>
-        )}
-        {getResults.isSuccess && <CCallout color="info">{getResults.data?.Results}</CCallout>}
-        {getResults.isError && (
-          <CCallout color="danger">Could not connect to API: {getResults.error.message}</CCallout>
-        )}
-        <CippDatatable
-          keyField="id"
-          reportName={`${tenant?.defaultDomainName}-MEMPolicyTemplates-List`}
-          path="/api/ListIntuneTemplates"
-          columns={columns}
-          params={{ TenantFilter: tenant?.defaultDomainName }}
-        />
-      </CCardBody>
+      <CCard className="content-card">
+        <CCardHeader className="d-flex justify-content-between align-items-center">
+          <CCardTitle>Endpoint Manager Templates</CCardTitle>
+        </CCardHeader>
+        <CCardBody>
+          {getResults.isFetching && (
+            <CCallout color="info">
+              <CSpinner>Loading</CSpinner>
+            </CCallout>
+          )}
+          {getResults.isSuccess && <CCallout color="info">{getResults.data?.Results}</CCallout>}
+          {getResults.isError && (
+            <CCallout color="danger">Could not connect to API: {getResults.error.message}</CCallout>
+          )}
+
+          <CippDatatable
+            keyField="id"
+            reportName={`${tenant?.defaultDomainName}-MEMPolicyTemplates-List`}
+            path="/api/ListIntuneTemplates"
+            columns={columns}
+            params={{ TenantFilter: tenant?.defaultDomainName }}
+          />
+        </CCardBody>
+      </CCard>
     </CippPage>
   )
 }
